Fix svg imports following VesselSidebar move
