+++ frontend/src/features/Vessel/components/VesselSidebar/Equipment/resume/BeaconMalfunctionCard.tsx
@@ -10,8 +10,8 @@ import {
 import { setOpenedBeaconMalfunction } from '../../../../../../domain/shared_slices/BeaconMalfunction'
 import { useMainAppDispatch } from '../../../../../../hooks/useMainAppDispatch'
 import { getDateTime } from '../../../../../../utils'
-import AtSeaSVG from '../../../icons/Icone_avarie_mer.svg?react'
-import AtPortSVG from '../../../icons/Icone_avarie_quai.svg?react'
+import AtSeaSVG from '../../../../../icons/Icone_avarie_mer.svg?react'
+import AtPortSVG from '../../../../../icons/Icone_avarie_quai.svg?react'
 
 import type { BeaconMalfunctionResumeAndDetails } from '../../../../../../domain/entities/beaconMalfunction/types'
 import type { HTMLProps } from 'react'
