+++ frontend/src/features/Vessel/components/VesselSidebar/Summary.tsx
@@ -1,19 +1,19 @@
+import { COLORS } from '@constants/constants'
 import { useMainAppDispatch } from '@hooks/useMainAppDispatch'
 import { useMainAppSelector } from '@hooks/useMainAppSelector'
+import { useIsSuperUser } from 'auth/hooks/useIsSuperUser'
+import { getCoordinates } from 'coordinates'
+import { WSG84_PROJECTION } from 'domain/entities/map/constants'
+import { showVessel } from 'domain/use_cases/vessel/showVessel'
 import { useEffect, useState } from 'react'
 import { FingerprintSpinner } from 'react-epic-spinners'
 import styled from 'styled-components'
 import * as timeago from 'timeago.js'
+import { getDateTime, timeagoFrenchLocale } from 'utils'
 
 import { RiskFactorResume } from './risk_factor/RiskFactorResume'
-import { useIsSuperUser } from '../../../../auth/hooks/useIsSuperUser'
-import { COLORS } from '../../../../constants/constants'
-import { getCoordinates } from '../../../../coordinates'
-import { WSG84_PROJECTION } from '../../../../domain/entities/map/constants'
-import { showVessel } from '../../../../domain/use_cases/vessel/showVessel'
-import { getDateTime, timeagoFrenchLocale } from '../../../../utils'
-import InfoSVG from '../../icons/Information.svg?react'
-import NoVesselSVG from '../../icons/Picto_photo_navire_manquante.svg?react'
+import InfoSVG from '../../../icons/Information.svg?react'
+import NoVesselSVG from '../../../icons/Picto_photo_navire_manquante.svg?react'
 
 // @ts-ignore
 timeago.register('fr', timeagoFrenchLocale)
