+++ frontend/src/features/Vessel/components/VesselSidebar/Tabs.tsx
@@ -1,21 +1,21 @@
+import { COLORS } from '@constants/constants'
+import { useMainAppDispatch } from '@hooks/useMainAppDispatch'
+import { useMainAppSelector } from '@hooks/useMainAppSelector'
 import { DisplayedErrorKey } from '@libs/DisplayedError/constants'
+import { useIsSuperUser } from 'auth/hooks/useIsSuperUser'
+import { forbiddenVesselSidebarPaths } from 'domain/entities/authorization/constants'
+import { VesselSidebarTab } from 'domain/entities/vessel/vessel'
+import { displayedErrorActions } from 'domain/shared_slices/DisplayedError'
+import { showVesselSidebarTab } from 'domain/shared_slices/Vessel'
 import { useEffect } from 'react'
 import styled from 'styled-components'
 
-import { useIsSuperUser } from '../../../../auth/hooks/useIsSuperUser'
-import { COLORS } from '../../../../constants/constants'
-import { forbiddenVesselSidebarPaths } from '../../../../domain/entities/authorization/constants'
-import { VesselSidebarTab } from '../../../../domain/entities/vessel/vessel'
-import { displayedErrorActions } from '../../../../domain/shared_slices/DisplayedError'
-import { showVesselSidebarTab } from '../../../../domain/shared_slices/Vessel'
-import { useMainAppDispatch } from '../../../../hooks/useMainAppDispatch'
-import { useMainAppSelector } from '../../../../hooks/useMainAppSelector'
-import ReportingSVG from '../icons/Icone_onglet_signalement.svg?react'
-import VMSSVG from '../icons/Icone_VMS_fiche_navire.svg?react'
-import ControlsSVG from '../icons/Picto_controles.svg?react'
-import VesselIDSVG from '../icons/Picto_identite.svg?react'
-import FisheriesSVG from '../icons/Picto_peche.svg?react'
-import SummarySVG from '../icons/Picto_resume.svg?react'
+import ReportingSVG from '../../../icons/Icone_onglet_signalement.svg?react'
+import VMSSVG from '../../../icons/Icone_VMS_fiche_navire.svg?react'
+import ControlsSVG from '../../../icons/Picto_controles.svg?react'
+import VesselIDSVG from '../../../icons/Picto_identite.svg?react'
+import FisheriesSVG from '../../../icons/Picto_peche.svg?react'
+import SummarySVG from '../../../icons/Picto_resume.svg?react'
 
 // TODO Move the icons in Monitor UI : https://github.com/MTES-MCT/monitorfish/issues/1736
 export function Tabs() {
