+++ frontend/src/features/Vessel/components/VesselSidebar/risk_factor/RiskFactorExplanationModal.tsx
@@ -11,9 +11,9 @@ import {
 } from '../../../../../domain/entities/vessel/riskFactor'
 import { StyledModalHeader } from '../../../../commonComponents/StyledModalHeader'
 import { basePrimaryButton, SecondaryButton } from '../../../../commonStyles/Buttons.style'
-import RiskFactorControlSVG from '../../icons/Note_de_controle_gyrophare.svg?react'
-import RiskFactorImpactSVG from '../../icons/Note_impact_poisson.svg?react'
-import RiskFactorInfractionsSVG from '../../icons/Note_infraction_stop.svg?react'
+import RiskFactorControlSVG from '../../../../icons/Note_de_controle_gyrophare.svg?react'
+import RiskFactorImpactSVG from '../../../../icons/Note_impact_poisson.svg?react'
+import RiskFactorInfractionsSVG from '../../../../icons/Note_infraction_stop.svg?react'
 
 import type { Promisable } from 'type-fest'
