+++ frontend/src/features/Vessel/components/VesselSidebar/risk_factor/RiskFactorExplanationSchema.tsx
@@ -2,9 +2,9 @@ import styled from 'styled-components'
 
 import { RiskFactorBox } from './RiskFactorBox'
 import { COLORS } from '../../../../../constants/constants'
-import RiskFactorControlSVG from '../../icons/Note_de_controle_gyrophare.svg?react'
-import RiskFactorImpactSVG from '../../icons/Note_impact_poisson.svg?react'
-import RiskFactorInfractionsSVG from '../../icons/Note_infraction_stop.svg?react'
+import RiskFactorControlSVG from '../../../../icons/Note_de_controle_gyrophare.svg?react'
+import RiskFactorImpactSVG from '../../../../icons/Note_impact_poisson.svg?react'
+import RiskFactorInfractionsSVG from '../../../../icons/Note_infraction_stop.svg?react'
 
 export function RiskFactorExplanationSchema() {
   return (
